split project-related e2e tests into smaller subsets
This was long overdue - I should have done this while splitting the
sprawling blob of monolithing e2e runs into subsets, but well, better
now than never. Project-related tests within API e2e tests were among
the longest-running, and the split by concern in this commit should
make each subset more manageable (for example, leading to shorter run
times if needing to trigger a new run of a subset manually because of
spurious failures in CI).
