+++ .github/workflows/tests-api-e2e.yml
@@ -28,7 +28,15 @@ jobs:
           - 'implicit-permissions'
           - 'integration'
           - 'legacy-project-import'
-          - 'project'
+          - 'projects/blm-calibration'
+          - 'projects/crud'
+          - 'projects/project-feature-tags'
+          - 'projects/user-projects'
+          - 'projects/cost-surfaces'
+          - 'projects/project-cloning'
+          - 'projects/project-scenario-comparison'
+          - 'projects/project-summaries'
+          - 'projects/published-projects'
           - 'project-jobs-status'
           - 'project-planning-areas-tiles'
           - 'project-planning-units'
