+++ api/apps/api/test/projects/crud/delete-project.e2e-spec.ts
