+++ api/apps/api/test/projects/crud/project-get.e2e-spec.ts
@@ -1,5 +1,5 @@
 import { FixtureType } from '@marxan/utils/tests/fixture-type';
-import { getFixtures } from './projects.fixtures';
+import { getFixtures } from '../projects.fixtures';
 
 let fixtures: FixtureType<typeof getFixtures>;
