+++ api/apps/api/test/projects/crud/update-project.fixtures.ts
@@ -1,11 +1,11 @@
-import { bootstrapApplication } from '../utils/api-application';
-import { GivenUserIsLoggedIn } from '../steps/given-user-is-logged-in';
-import { OrganizationsTestUtils } from '../utils/organizations.test.utils';
-import { E2E_CONFIG } from '../e2e.config';
-import { ProjectsTestUtils } from '../utils/projects.test.utils';
+import { bootstrapApplication } from '../../utils/api-application';
+import { GivenUserIsLoggedIn } from '../../steps/given-user-is-logged-in';
+import { OrganizationsTestUtils } from '../../utils/organizations.test.utils';
+import { E2E_CONFIG } from '../../e2e.config';
+import { ProjectsTestUtils } from '../../utils/projects.test.utils';
 import * as request from 'supertest';
 import { ProjectChecker } from '@marxan-api/modules/projects/project-checker/project-checker.service';
-import { ProjectCheckerFake } from '../utils/project-checker.service-fake';
+import { ProjectCheckerFake } from '../../utils/project-checker.service-fake';
 import { HttpStatus } from '@nestjs/common';
 
 export const getFixtures = async () => {
