+++ api/apps/api/test/projects/project-summaries/project-output-summary.e2e-spec.ts
@@ -1,6 +1,6 @@
-import { bootstrapApplication } from '../utils/api-application';
-import { GivenUserIsLoggedIn, userObj } from '../steps/given-user-is-logged-in';
-import { GivenProjectExists } from '../steps/given-project';
+import { bootstrapApplication } from '../../utils/api-application';
+import { GivenUserIsLoggedIn, userObj } from '../../steps/given-user-is-logged-in';
+import { GivenProjectExists } from '../../steps/given-project';
 import { Repository } from 'typeorm';
 import { getRepositoryToken } from '@nestjs/typeorm';
 import * as archiver from 'archiver';
